requirements.txt
torch==2.0.1
torchvision==0.15.2
pytorch_lightning==2.0.7
Pillow
opencv-python
transformers
taming-transformers-rom1504 
tqdm
numpy
kornia
webdataset
omegaconf
einops
scikit-image
pymcubes
carvekit-colab
open3d
trimesh
easydict
nerfacc
raymarching
imageio-ffmpeg==0.4.7
git+https://github.com/openai/CLIP.git
